fix crash at launch from popup import cycle
dcf0ff4 had lib/popup.ts and lib/window.ts import loadUrl from
lib/web-contents.ts, which imports @/ipc, which imports downloads.ts,
which imports lib/popup.ts. Downloads instantiates a Popup at module
scope, so the bundle evaluated it before the Popup class existed:
"TypeError: Popup is not a constructor".

Move loadUrl into lib/load-url.ts, a module with no app-level imports.
