+++ packages/app/lib/web-contents.ts
@@ -4,28 +4,10 @@ import {
   WebContentsView,
   type WebContentsViewConstructorOptions,
 } from "electron";
-import { serializeError } from "serialize-error";
 import { setupWindowContextMenu } from "@/context-menu";
 import { ipc } from "@/ipc";
 import { shouldOpenDevToolsOnLaunch } from "./dev-tools";
 import { log } from "./log";
-
-/**
- * `loadURL` rejects when the load never commits — a renderer that went away
- * mid-navigation, a network stack that gave up — and nothing waits on the
- * promise, which turns every such load into an unhandled rejection. Says what
- * happened and answers whether the page arrived.
- */
-export function loadUrl(webContents: WebContents, url: string) {
-  return webContents
-    .loadURL(url)
-    .then(() => true)
-    .catch((error: unknown) => {
-      log.error("Failed to load URL", { url, error: serializeError(error) });
-
-      return false;
-    });
-}
 
 /**
  * Says why a view is empty. A load that never arrives leaves nothing behind on
